USe color palette in sheet builder
