+++ src/webapp/logic/sheetBuilder.js
@@ -1,6 +1,6 @@
 import * as Excel from "excel4node";
 import _ from "lodash";
-import { baseStyle, createColumn, groupStyle, protectedSheet } from "../utils/excel";
+import { defaultColorScale } from "../utils/colors";
 import { buildAllPossiblePeriods } from "../utils/periods";
 import { getObjectVersion } from "./utils";
 
@@ -268,8 +268,7 @@ SheetBuilder.prototype.fillDataEntrySheet = function () {
     let columnId = 1;
     let groupId = 0;
 
-    createColumn(
-        this.workbook,
+    this.createColumn(
         dataEntrySheet,
         itemRow,
         columnId++,
@@ -278,11 +277,10 @@ SheetBuilder.prototype.fillDataEntrySheet = function () {
         this.validations.get("organisationUnits")
     );
     if (element.type === "program") {
-        createColumn(this.workbook, dataEntrySheet, itemRow, columnId++, "Latitude");
-        createColumn(this.workbook, dataEntrySheet, itemRow, columnId++, "Longitude");
+        this.createColumn(dataEntrySheet, itemRow, columnId++, "Latitude");
+        this.createColumn(dataEntrySheet, itemRow, columnId++, "Longitude");
     } else if (element.type === "dataSet") {
-        createColumn(
-            this.workbook,
+        this.createColumn(
             dataEntrySheet,
             itemRow,
             columnId++,
@@ -295,8 +293,7 @@ SheetBuilder.prototype.fillDataEntrySheet = function () {
     const { code: attributeCode } = metadata.get(element.categoryCombo?.id);
     const optionsTitle = attributeCode !== "default" ? `_${element.categoryCombo.id}` : "Options";
 
-    createColumn(
-        this.workbook,
+    this.createColumn(
         dataEntrySheet,
         itemRow,
         columnId++,
@@ -337,8 +334,7 @@ SheetBuilder.prototype.fillDataEntrySheet = function () {
                             const validation = dataElement.optionSet
                                 ? dataElement.optionSet.id
                                 : dataElement.valueType;
-                            createColumn(
-                                this.workbook,
+                            this.createColumn(
                                 dataEntrySheet,
                                 itemRow,
                                 columnId,
@@ -357,7 +353,7 @@ SheetBuilder.prototype.fillDataEntrySheet = function () {
                         dataEntrySheet
                             .cell(sectionRow, firstColumnId, sectionRow, columnId - 1, true)
                             .formula("_" + dataElement.id)
-                            .style(groupStyle(groupId));
+                            .style(this.groupStyle(groupId));
 
                         if (description !== undefined) {
                             dataEntrySheet
@@ -376,8 +372,7 @@ SheetBuilder.prototype.fillDataEntrySheet = function () {
         _.forEach(element.programStages, programStageT => {
             const programStage = metadata.get(programStageT.id);
 
-            createColumn(
-                this.workbook,
+            this.createColumn(
                 dataEntrySheet,
                 itemRow,
                 columnId++,
@@ -409,12 +404,11 @@ SheetBuilder.prototype.fillDataEntrySheet = function () {
                     const validation = dataElement.optionSet
                         ? dataElement.optionSet.id
                         : dataElement.valueType;
-                    createColumn(
-                        this.workbook,
+                    this.createColumn(
                         dataEntrySheet,
                         itemRow,
                         columnId,
-                        "_" + dataElement.id,
+                        `_${dataElement.id}`,
                         groupId,
                         this.validations.get(validation)
                     );
@@ -433,8 +427,8 @@ SheetBuilder.prototype.fillDataEntrySheet = function () {
                 if (firstColumnId < columnId)
                     dataEntrySheet
                         .cell(sectionRow, firstColumnId, sectionRow, columnId - 1, true)
-                        .formula("_" + programStageSection.id)
-                        .style(groupStyle(groupId));
+                        .formula(`_${programStageSection.id}`)
+                        .style(this.groupStyle(groupId));
 
                 groupId++;
             });
@@ -480,4 +474,93 @@ SheetBuilder.prototype.translate = function (item) {
     } else {
         return { name, description };
     }
+};
+
+SheetBuilder.prototype.createColumn = function (
+    sheet,
+    rowId,
+    columnId,
+    label,
+    groupId = null,
+    validation = null
+) {
+    sheet.column(columnId).setWidth(20);
+    const cell = sheet.cell(rowId, columnId);
+    cell.style(groupId !== null ? this.groupStyle(groupId) : baseStyle);
+
+    if (label.startsWith("_")) cell.formula(label);
+    else cell.string(label);
+
+    if (validation !== null) {
+        const ref = `${Excel.getExcelAlpha(columnId)}${rowId + 1}:${Excel.getExcelAlpha(
+            columnId
+        )}1048576`;
+        sheet.addDataValidation({
+            type: "list",
+            allowBlank: true,
+            error: "Invalid choice was chosen",
+            errorStyle: "warning",
+            showDropDown: true,
+            sqref: ref,
+            formulas: [validation.toString()],
+        });
+
+        sheet.addConditionalFormattingRule(ref, {
+            type: "expression", // the conditional formatting type
+            priority: 1, // rule priority order (required)
+            formula:
+                "ISERROR(MATCH(" +
+                Excel.getExcelAlpha(columnId) +
+                +(rowId + 1) +
+                "," +
+                validation.toString().substr(1) +
+                ",0))", // formula that returns nonzero or 0
+            style: this.workbook.createStyle({
+                font: {
+                    bold: true,
+                    color: "FF0000",
+                },
+            }), // a style object containing styles to apply
+        });
+    }
+};
+
+SheetBuilder.prototype.groupStyle = function (groupId) {
+    const { palette = defaultColorScale } = this.builder;
+    return {
+        ...baseStyle,
+        fill: {
+            type: "pattern",
+            patternType: "solid",
+            fgColor: palette[groupId % palette.length],
+        },
+    };
+};
+
+/**
+ * Common cell style definition
+ * @type {{alignment: {horizontal: string, vertical: string, wrapText: boolean, shrinkToFit: boolean}}}
+ */
+const baseStyle = {
+    alignment: {
+        horizontal: "center",
+        vertical: "center",
+        wrapText: true,
+        shrinkToFit: true,
+    },
+    fill: {
+        type: "pattern",
+        patternType: "solid",
+        fgColor: "ffffff",
+    },
+};
+
+const protectedSheet = {
+    sheetProtection: {
+        sheet: true,
+        formatCells: false,
+        formatColumns: false,
+        formatRows: false,
+        password: "Wiscentd2019!",
+    },
 };
