+++ src/webapp/utils/colors.ts
@@ -1,3 +1,4 @@
+import _ from "lodash";
 import { colorScales, palettes } from "./palettes";
 
 // Returns a color brewer scale for a number of classes
@@ -7,7 +8,7 @@ export const getColorPalette = (scale: string, classes: number): string[] => {
 
 // Returns color scale name for a palette
 export const getColorScale = (palette: string[]) => {
-    return colorScales.find(name => palettes[name][palette.length] === palette);
+    return colorScales.find(name => _.isEqual(palettes[name][palette.length], palette));
 };
 
 export const defaultColorScale = getColorPalette("default", 9);
